refactor class names in inheritance tree
