+++ TpSolver.Tests/Parallel/CycleSearcherTests.cs
@@ -2,7 +2,7 @@ using TpSolver.BfsSearch;
 using TpSolver.Perturbation;
 using TpSolver.Shared;
 using TpSolver.Tests.Sequential.Utils;
-using static TpSolver.BfsSearch.VamBase;
+using static TpSolver.BfsSearch.Vam;
 
 namespace TpSolver.Tests.Parallel;
 
@@ -22,7 +22,7 @@ public class CycleSearcherTests
         );
         TransportProblem degenerousTp = TransportProblem.GenerateRandom(size, size, limits);
 
-        var expected = new Vam(degenerousTp).Search();
+        var expected = new VamSeq(degenerousTp).Search();
         AllocationMatrix actual = new(expected);
         int perturbCount = actual.NRows + actual.NCols - 1 - actual.Count(a => a.IsBasic);
