+++ TpSolver.Tests/Parallel/PotentialsCalculatorTests.cs
@@ -1,4 +1,3 @@
-using System.Security.Authentication.ExtendedProtection;
 using TpSolver.Shared;
 using TpSolver.Solver.Modi;
 using TpSolver.Tests.Sequential.Utils;
@@ -15,11 +14,11 @@ public class PotentialsCalculatorTests
     {
         TransportProblem tp = TransportProblem.GenerateRandom(size, size);
 
-        var seq = new ModiSolver(tp) { Profiler = new() };
+        var seq = new ModiSolverSeq(tp) { Profiler = new() };
         var par = new ModiSolverParallel(tp, 6) { Profiler = new() };
 
-        AllocationMatrix? expected = seq.Solve(tp);
-        AllocationMatrix? actual = par.Solve(tp);
+        AllocationMatrix? expected = seq.Solve();
+        AllocationMatrix? actual = par.Solve();
         Assert.NotNull(expected);
         Assert.NotNull(actual);
         Assert.Equal(expected.AsEnumerableNBDistinct(), actual.AsEnumerableNBDistinct());
